+++ packages/elements-react/graphing/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pie-element/graphing
 
+## 10.1.2-next.9
+
+### Patch Changes
+
+- Updated dependencies [7cae8f9]
+- Updated dependencies [a204d11]
+  - @pie-lib/render-ui@6.2.0-next.40
+  - @pie-lib/graphing@4.0.5-next.35
+  - @pie-lib/editable-html-tip-tap@2.1.2-next.34
+  - @pie-lib/config-ui@13.0.4-next.34
+  - @pie-lib/correct-answer-toggle@4.0.3-next.40
+
 ## 10.1.2-next.8
 
 ### Patch Changes