+++ packages/shared/theming-mui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pie-element/shared-theming-mui
 
+## 0.1.1-next.3
+
+### Patch Changes
+
+- Updated dependencies [7cae8f9]
+  - @pie-element/shared-theming@0.2.0-next.3
+
 ## 0.1.1-next.2
 
 ### Patch Changes